refactor dashboard
